requirements.txt
fastapi
uvicorn
python-docx
pyyaml
requests
